Rebase, better wording of testcase, linting changes for c51f5f95970cf0196064ea48478913cc1c17956a
