+++ lib/sinon/mock.js
@@ -45,18 +45,20 @@
             }
         }
 
-        sinon.extend(mock, {
-            function arrayEquals(arr1, arr2, compareLength) {
-                if (compareLength && (arr1.length !== arr2.length)) return false;
-                for (var i = 0, l = arr1.length; i < l; i++) {
-                    if (!sinon.deepEqual(arr1[i], arr2[i])) {
-                        return false;
-                    }
+        function arrayEquals(arr1, arr2, compareLength) {
+            if (compareLength && (arr1.length !== arr2.length)) {
+                return false;
+            }
+
+            for (var i = 0, l = arr1.length; i < l; i++) {
+                if (!sinon.deepEqual(arr1[i], arr2[i])) {
+                    return false;
                 }
-                return true;
+            }
+            return true;
         }
 
-        return {
+        sinon.extend(mock, {
             create: function create(object) {
                 if (!object) {
                     throw new TypeError("object is null");
