+++ test/mock-test.js
@@ -297,6 +297,20 @@
                         obj.foobar();
                         mock.verify();
                     });
+                },
+
+                "should not throw when exceeding at least expectation and withargs": function () {
+                    var obj = { foobar: function () {} };
+                    var mock = sinon.mock(obj);
+
+                    mock.expects("foobar").withArgs("arg1");
+                    mock.expects("foobar").atLeast(1).withArgs("arg2");
+
+                    obj.foobar("arg1");
+                    obj.foobar("arg2");
+                    obj.foobar("arg2");
+
+                    assert(mock.verify());
                 }
             },
